--- a/lib/core/network/entryPoint.js
+++ b/lib/core/network/entryPoint.js
@@ -25,10 +25,6 @@ const fs = require('fs');
 const path = require('path');
 
 const Bluebird = require('bluebird');
-const moment = require('moment');
-const winston = require('winston');
-const WinstonElasticsearch = require('winston-elasticsearch');
-const WinstonSyslog = require('winston-syslog');
 
 const { RequestContext } = require('../../api/request');
 const Context = require('./context');
@@ -37,9 +33,9 @@ const MqttProtocol = require('./protocols/mqttProtocol');
 const InternalProtocol = require('./protocols/internalProtocol');
 const HttpWsProtocol = require('./protocols/httpwsProtocol');
 const Manifest = require('./protocolManifest');
-const ClientConnection = require('./clientConnection');
 const removeErrorStack = require('./removeErrorStack');
 const kerror = require('../../kerror');
+const { AccessLogger } = require('./accessLogger');
 
 const networkError = kerror.wrap('network', 'entrypoint');
 
@@ -57,11 +53,9 @@ class EntryPoint {
 
     this._clients = new Map();
 
-    this.logger = null;
+    this.accessLogger = new AccessLogger();
 
     this.isShuttingDown = false;
-
-    this.anonymousUserId = null;
   }
 
   dispatch (event, data) {
@@ -109,7 +103,7 @@ class EntryPoint {
       throw networkError.get('invalid_port', this.config.port);
     }
 
-    this.initLogger();
+    await this.accessLogger.init();
 
     const initPromises = [];
 
@@ -130,8 +124,6 @@ class EntryPoint {
 
     await Bluebird.all(initPromises);
 
-    const anonymous = await global.kuzzle.ask('core:security:user:anonymous:get');
-    this.anonymousUserId = anonymous._id;
 
     await this.loadMoreProtocols();
   }
@@ -224,247 +216,13 @@ class EntryPoint {
     });
   }
 
-  initLogger () {
-    const transports = [];
-
-    for (const conf of this.config.logs.transports) {
-      const opts = {
-        colorize: conf.colorize === true
-          ? winston.format.colorize()
-          : false,
-        depth: conf.depth || false,
-        format: conf.format
-          ? winston.format[conf.format]()
-          : winston.format.json(),
-        level: conf.level || 'info',
-        prettyPrint: conf.prettyPrint === true
-          ? winston.format.prettyPrint()
-          : false,
-        silent: conf.silent || false,
-        timestamp: conf.timestamp === true
-          ? winston.format.timestamp()
-          : false
-      };
-
-      switch (conf.transport || 'console') {
-        case 'console':
-          transports.push(
-            new winston.transports.Console(
-              Object.assign(
-                opts,
-                {
-                  humanReadableUnhandledException: conf.humanReadableUnhandledException || true,
-                  stderrLevels: conf.stderrLevels || ['error', 'debug']
-                })));
-          break;
-        case 'elasticsearch':
-          transports.push(
-            new WinstonElasticsearch(
-              Object.assign(
-                opts,
-                {
-                  clientOpts: conf.clientOpts || {},
-                  ensureMappingTemplate: conf.ensureMappingTemplate !== false,
-                  flushInterval: conf.flushInterval || 2000,
-                  index: conf.index,
-                  indexPrefix: conf.indexPrefix || 'kuzzle-access',
-                  indexSuffixPattern: conf.indexSuffixPattern || 'YYYY.MM',
-                  mappingTemplate: conf.mappingTemplate || 'access.log.mapping.json',
-                  messageType: conf.messageType || 'access'
-                })));
-          break;
-        case 'file':
-          transports.push(
-            new winston.transports.File(
-              Object.assign(
-                opts,
-                {
-                  eol: conf.eol || '\n',
-                  filename: conf.filename || 'kuzzle.access.log',
-                  logstash: conf.logstash || false,
-                  maxFiles: conf.maxFiles,
-                  maxRetries: conf.maxRetries || 2,
-                  maxSize: conf.maxSize,
-                  tailable: conf.tailable,
-                  zippedArchive: conf.zippedArchive || false
-                })));
-          break;
-        case 'syslog':
-          transports.push(
-            new WinstonSyslog(
-              Object.assign(
-                opts,
-                {
-                  app_name: conf.app_name || process.title,
-                  eol: conf.eol,
-                  facility: conf.facility || 'local0',
-                  host: conf.host || 'localhost',
-                  localhost: conf.localhost || 'localhost',
-                  path: conf.path || '/dev/log',
-                  pid: conf.pid || process.pid,
-                  port: conf.port || 514,
-                  protocol: conf.protocol || 'udp4',
-                  type: conf.type || 'BSD'
-                })));
-          break;
-        default:
-          global.kuzzle.log.error(`Failed to initialize logger transport "${conf.transport}": unsupported transport. Skipped.`);
-      }
-    }
-
-    this.logger = winston.createLogger({transports});
-  }
-
   /**
-   * /!\ This method is a critical section of code.
-   * Modifications must be compared against current
-   * performances benchmarks
-   *
+   * @param {ClientConnection} connection
    * @param {Request} request
    * @param {object} extra
    */
-  logAccess (request, extra = null) {
-    let connection = this._clients.get(request.context.connection.id);
-
-    // Make do with the RequestContext information
-    if (!connection) {
-      connection = new ClientConnection(
-        request.context.connection.protocol || '-',
-        request.context.connection.ips || [],
-        request.context.connection.misc.headers);
-    }
-
-    if (this.config.logs.accessLogFormat === 'logstash') {
-      // custom kuzzle logs to be exported to logstash
-      this.logger.info({
-        connection,
-        error: request.error,
-        extra,
-        request: request.input,
-        status: request.status
-      });
-      return;
-    }
-
-    // user init: prioritize the already decoded and verified token stored in
-    // the request
-    // If not available, then that may mean that we didn't verify the user yet,
-    // so we have to decode any provided token
-    let user = null;
-
-    if (request.context.token !== null) {
-      user = request.context.token.userId === this.anonymousUserId
-        ? '(anonymous)'
-        : request.context.token.userId;
-    }
-
-    // = apache combined
-    const protocol = connection.protocol.toUpperCase();
-    let
-      url,
-      verb = 'DO';
-
-    if (connection.protocol.indexOf('HTTP/') === 0) {
-      verb = extra.method;
-      url = extra.url;
-
-      // try to get plain user name, 1st form jwt token if present, then from
-      // basic auth
-      const authHdr = connection.headers.authorization;
-      if (user === null && authHdr) {
-        try {
-          if (authHdr.toLowerCase().indexOf('bearer') === 0) {
-            const
-              b64Payload = authHdr.split('.')[1],
-              payload = Buffer.from(b64Payload, 'base64').toString('utf8');
-
-            user = JSON.parse(payload)._id;
-          }
-          else {
-            user = Buffer
-              .from(authHdr, 'base64')
-              .toString('utf8')
-              .split(':')[0];
-          }
-        }
-        catch (err) {
-          // do nothing: we don't know anything about the token, it may be
-          // invalid on-purpose for all we know
-        }
-      }
-    }
-    // for other protocols than http, we rebuild a pseudo url
-    else {
-      url = `/${request.input.controller}/${request.input.action}`;
-
-      if (request.input.args.index) {
-        url += `/${request.input.args.index}`;
-      }
-
-      if (request.input.args.collection) {
-        url += `/${request.input.args.collection}`;
-      }
-
-      if (request.input.args._id) {
-        url += `/${request.input.args._id}`;
-      }
-
-      let queryString = '';
-
-      for (const k of Object.keys(request.input.args)) {
-        if (k === '_id' || k === 'index' || k === 'collection') {
-          continue;
-        }
-
-        const val = request.input.args[k];
-
-        if (queryString.length > 0) {
-          queryString += '&';
-        }
-
-        queryString += `${k}=${typeof val === 'object' ? JSON.stringify(val) : val}`;
-      }
-
-      if (queryString.length > 0) {
-        url += `?${queryString}`;
-      }
-
-      if (user === null && request.input.jwt) {
-        try {
-          const
-            b64Payload = request.input.jwt.split('.')[1],
-            payload = Buffer.from(b64Payload, 'base64').toString('utf8');
-
-          user = JSON.parse(payload)._id;
-        }
-        catch (err) {
-          // do nothing: we don't know anything about the token, it may be
-          // invalid on-purpose for all we know
-        }
-      }
-    }
-
-    if (user === null) {
-      user = '-';
-    }
-
-    const
-      ip = connection.ips[connection.ips.length - 1 - this.config.logs.accessLogIpOffset]
-        || '-',
-      ts = moment().format('DD/MMM/YYYY:HH:mm:ss ZZ'),
-      status = request.status || '-',
-      size = request.response
-        ? Buffer.byteLength(JSON.stringify(request.response))
-        : '-',
-      referer = connection.headers.referer
-        ? `"${connection.headers.referer}"`
-        : '-',
-      agent = connection.headers['user-agent']
-        ? `"${connection.headers['user-agent']}"`
-        : '-';
-
-
-    this.logger.info(`${ip} - ${user} [${ts}] "${verb} ${url} ${protocol}" ${status} ${size} ${referer} ${agent}`);
+  logAccess (connection, request, extra = null) {
+    this.accessLogger.log(connection, request, extra);
   }
 
   /*
@@ -474,14 +232,14 @@ class EntryPoint {
   */
 
   /**
-   *
+   * @param {ClientConnection} connection
    * @param {Request} request
    * @param cb
    */
-  execute (request, cb) {
+  execute (connection, request, cb) {
     if (this.isShuttingDown) {
       debug('Shutting down. Dropping request: %a', request);
-      this._isShuttingDownError(request, cb);
+      this._isShuttingDownError(connection, request, cb);
       return;
     }
 
@@ -494,7 +252,7 @@ class EntryPoint {
         _res.setError(error);
       }
 
-      this.logAccess(_res);
+      this.logAccess(connection, _res);
 
       const response = _res.response.toJSON();
 
@@ -554,9 +312,9 @@ class EntryPoint {
     }
   }
 
-  _isShuttingDownError (request, cb) {
+  _isShuttingDownError (connection, request, cb) {
     request.setError(networkError.get('shutting_down'));
-    this.logAccess(request);
+    this.logAccess(connection, request);
 
     cb(removeErrorStack(request.response.toJSON()));
   }