--- a/lib/core/network/protocols/httpwsProtocol.js
+++ b/lib/core/network/protocols/httpwsProtocol.js
@@ -333,7 +333,7 @@ class HttpWsProtocol extends Protocol {
       return;
     }
 
-    this.entryPoint.execute(request, result => {
+    this.entryPoint.execute(connection, request, result => {
       if (result.content) {
         if (typeof result.content === 'object') {
           result.content.room = result.requestId;
@@ -558,7 +558,7 @@ class HttpWsProtocol extends Protocol {
     this.entryPoint.newConnection(message.connection);
 
     global.kuzzle.router.http.route(message, request => {
-      this.entryPoint.logAccess(request, message);
+      this.entryPoint.logAccess(message.connection, request, message);
       this.entryPoint.removeConnection(message.connection.id);
 
       if (response.aborted) {
@@ -624,6 +624,7 @@ class HttpWsProtocol extends Protocol {
     debugHTTP('[%s] httpSendError: %a', message.connection.id, kerr);
 
     this.entryPoint.logAccess(
+      message.connection,
       new Request(message, {
         connectionId: message.connection.id,
         error: kerr,