--- a/lib/core/network/protocols/mqttProtocol.js
+++ b/lib/core/network/protocols/mqttProtocol.js
@@ -202,6 +202,7 @@ class MqttProtocol extends Protocol {
       const request = new Request(payload, { connection });
 
       this.entryPoint.execute(
+        connection,
         request,
         response => this._respond(client, response));
     }