--- a/lib/model/security/role.js
+++ b/lib/model/security/role.js
@@ -21,7 +21,6 @@
 
 'use strict';
 
-const { isNil } = require('lodash');
 const kerror = require('../../kerror');
 const { has, isPlainObject } = require('../../util/safeObject');
 
@@ -45,7 +44,7 @@ class Role {
       throw kerror.get('security', 'role', 'uninitialized', this._id);
     }
 
-    if (isNil(this.controllers)) {
+    if (this.controllers === undefined || this.controllers === null) {
       return false;
     }
 
@@ -96,7 +95,7 @@ class Role {
    * @returns {Promise}
    */
   async validateDefinition () {
-    if (isNil(this.controllers)) {
+    if (this.controllers === undefined || this.controllers === null) {
       throw assertionError.get('missing_argument', 'controllers');
     }
 
@@ -193,7 +192,7 @@ function checkIndexRestriction(request, restriction) {
     return true;
   }
 
-  return restriction.collections.indexOf(request.input.args.collection) > -1;
+  return restriction.collections.includes(request.input.args.collection);
 }
 
 /**
@@ -207,7 +206,8 @@ function checkRestrictions(request, restrictedTo) {
     return true;
   }
 
-  // If the request's action does not refer to an index, restrictions are useless for this action (=> ignore them):
+  // If the request's action does not refer to an index, restrictions are
+  // useless for this action (=> ignore them)
   if (!request.input.args.index) {
     return true;
   }