--- a/package.json
+++ b/package.json
@@ -72,6 +72,7 @@
     "moment": "^2.29.1",
     "ms": "^2.1.3",
     "murmurhash-native": "^3.5.0",
+    "nanoid": "^3.1.20",
     "passport": "^0.4.1",
     "protobufjs": "~6.10.2",
     "rc": "1.2.8",