… test/core/network/accessLogger.test.js
… test/core/network/entryPoint.test.js