@scottinet 2021-05-04
@rolljee > UUIDs are still used as the default document ID generator. As NanoIds have a different form, I prefer to avoid changing the way our default IDs look.

With this new version, UUID is rarely used, so this shouldn't have an impact on performances.